WIP作成していたメソッドをクラスごとにまとめなおした
